+++ xpasses/XR001/testdata/src/a/outside_package.go
@@ -1,0 +1,11 @@
+package a
+
+import (
+	"a/schema"
+)
+
+func foutside() {
+	var d schema.ResourceData
+
+	d.GetOkExists("test")
+}