+++ xpasses/XR003/testdata/src/a/comment_ignore.go
@@ -1,0 +1,12 @@
+package a
+
+import (
+	"github.com/hashicorp/terraform-plugin-sdk/helper/schema"
+)
+
+func fcommentignore() {
+	//lintignore:XR003
+	_ = schema.Resource{
+		Create: createFunc,
+	}
+}