+++ xpasses/XR003/testdata/src/a/outside_package.go
@@ -1,0 +1,15 @@
+package a
+
+import (
+	"a/schema"
+)
+
+func foutside() {
+	_ = schema.Resource{
+		Create: outsideCreateFunc,
+	}
+}
+
+func outsideCreateFunc(d *schema.ResourceData, meta interface{}) error {
+	return nil
+}