+++ xpasses/XR003/testdata/src/a/schema/schema.go
@@ -1,0 +1,9 @@
+package schema
+
+type Resource struct {
+	Create CreateFunc
+}
+
+type ResourceData struct{}
+
+type CreateFunc func(*ResourceData, interface{}) error